+++ src/io/io_mod.f90
@@ -134,7 +134,7 @@ module io
       integer(int64), dimension(1), intent(in) :: global_start !< What global index to start reading from
       integer(int64), dimension(1), intent(in) :: count        !< How many array element to read
       integer(int32), dimension(:), intent(inout) :: var       !< The 1D integer array
-      integer(int32), optional, intent(in) :: step             !< The step to read
+      integer(int64), optional, intent(in) :: step             !< The step to read
     end subroutine
 
     !> Read a 1D 64-bit integer array from file
@@ -154,7 +154,7 @@ module io
       integer(int64), dimension(2), intent(in) :: global_start !< What global index to start reading from
       integer(int64), dimension(2), intent(in) :: count        !< How many array element to read
       integer(int32), dimension(:, :), intent(inout) :: var    !< The 2D integer array
-      integer(int32), optional, intent(in) :: step             !< The step to read
+      integer(int64), optional, intent(in) :: step             !< The step to read
     end subroutine
 
     !> Read a 2D 64-bit integer array from file
@@ -164,7 +164,7 @@ module io
       integer(int64), dimension(2), intent(in) :: global_start !< What global index to start reading from
       integer(int64), dimension(2), intent(in) :: count        !< How many array element to read
       integer(int64), dimension(:, :), intent(inout) :: var    !< The 2D integer array
-      integer(int32), optional, intent(in) :: step             !< The step to read
+      integer(int64), optional, intent(in) :: step             !< The step to read
     end subroutine
 
     !> Read a 1D 32-bit real array from file
@@ -174,7 +174,7 @@ module io
       integer(int64), dimension(1), intent(in) :: global_start !< What global index to start reading from
       integer(int64), dimension(1), intent(in) :: count        !< How many array element to read
       real(real32), dimension(:), intent(inout) :: var         !< The 1D real array
-      integer(int32), optional, intent(in) :: step             !< The step to read
+      integer(int64), optional, intent(in) :: step             !< The step to read
     end subroutine
 
     !> Read a 1D 64-bit real array from file
@@ -194,7 +194,7 @@ module io
       integer(int64), dimension(2), intent(in) :: global_start !< What global index to start reading from
       integer(int64), dimension(2), intent(in) :: count        !< How many array element to read
       real(real32), dimension(:, :), intent(inout) :: var      !< The 2D real array
-      integer(int32), optional, intent(in) :: step             !< The step to read
+      integer(int64), optional, intent(in) :: step             !< The step to read
     end subroutine
 
     !> Read a 2D 64-bit real array from file
